get a draft for the phoenix agent
