+++ games/phoenix/phoenix_agent.py
@@ -1,65 +1,85 @@
 import time
 import numpy as np
-import concurrent.futures
 import argparse
-from collections import deque
+from collections import deque, Counter
+import shutil
 
-from games.phoenix.workers import phoenix_worker
+import os
+import json
+import re
+import pyautogui
+
+from games.ace_attorney.workers import ace_attorney_worker, perform_move
 from tools.utils import str2bool
-# System prompt remains constant
-system_prompt = (  
-    "You are a highly intelligent Phoenix Wright: Ace Attorney gameplay agent trained to achieve the highest possible score "  
-    "using a limited number of moves. Your goal is to analyze the board carefully and identify the best adjacent " 
-) 
+from collections import Counter
 
+CACHE_DIR = "cache/ace_attorney"
+
+# System prompt remains constant
+system_prompt = (
+    "You are an expert AI agent specialized in playing Ace Attorney games. Your goal is to solve cases by gathering evidence, cross-examining witnesses, and presenting the correct evidence at the right time to prove your client's innocence. "
+)
 
 def main():
-    """
-    Spawns a number of short-term and/or long-term workers based on user-defined parameters.
-    """
-    parser = argparse.ArgumentParser(
-        description="Phoenix Wright: Ace Attorney grid processing with configurable parameters."
-    )
-    parser.add_argument("--api_provider", type=str, default="openai",
-                        help="API provider to use (anthropic, openai, gemini).")
-    parser.add_argument("--model_name", type=str, default="gpt-4-turbo",
-                        help="Model name.")
-    parser.add_argument("--modality", type=str, default="text-only", choices=["text-only", "vision-text"], help="modality used.")
-    parser.add_argument("--thinking", type=str, default=True, help="Whether to use deep thinking.")
-    parser.add_argument("--crop_left", type=int, default=0, help="Pixels to crop from the left.")
-    parser.add_argument("--crop_right", type=int, default=730, help="Pixels to crop from the right.")
-    parser.add_argument("--crop_top", type=int, default=230, help="Pixels to crop from the top.")
-    parser.add_argument("--crop_bottom", type=int, default=230, help="Pixels to crop from the bottom.")
-    parser.add_argument("--grid_rows", type=int, default=7, help="Number of grid rows.")
-    parser.add_argument("--grid_cols", type=int, default=7, help="Number of grid columns.")
-    
+    # Delete existing cache directory if it exists and create a new one
+    if os.path.exists(CACHE_DIR):
+        shutil.rmtree(CACHE_DIR)
+    os.makedirs(CACHE_DIR, exist_ok=True)
+
+    parser = argparse.ArgumentParser(description="Ace Attorney AI Agent")
+    parser.add_argument("--api_provider", type=str, default="anthropic", help="API provider to use.")
+    parser.add_argument("--model_name", type=str, default="claude-3-7-sonnet-20250219", help="LLM model name.")
+    parser.add_argument("--modality", type=str, default="vision-text", choices=["text-only", "vision-text"],
+                        help="modality used.")
+    parser.add_argument("--thinking", type=str, default="True", help="Whether to use deep thinking.")
+    parser.add_argument("--episode_name", type=str, default="The First Turnabout", 
+                        help="Name of the current episode being played.")
     args = parser.parse_args()
 
-    prev_responses = deque(maxlen=7)
-    evidence = False
+    prev_response = ""
+
     try:
         while True:
             start_time = time.time()
-            # Execute the phoenix worker
-            latest_response, key = phoenix_worker(system_prompt, args.api_provider, args.model_name, args.modality, str2bool(args.thinking),
-                                                 args.crop_left, args.crop_right, args.crop_top, args.crop_bottom, 
-                                                 args.grid_rows, args.grid_cols, " ".join(prev_responses), evidence)
-            # break
-            if latest_response:
-                prev_responses.append(latest_response)
+
+            # Convert thinking string to boolean
+            thinking_bool = str2bool(args.thinking)
+
+            # Direct call to ace_attorney_worker
+            result = ace_attorney_worker(
+                system_prompt,
+                args.api_provider,
+                args.model_name,
+                prev_response,
+                thinking=thinking_bool,
+                modality=args.modality,
+                episode_name=args.episode_name
+            )
+            
+
+            # Process the result
+            if result:
+                # Extract game state, move and thought from the result
+                game_state = result["game_state"]
+                move = result["move"].strip().lower()
+                thought = result["thought"]
+                
+                # Print current game state
+                print(f"Current Game State: {game_state}")
+                
+                # Perform the move using the imported function
+                perform_move(move)
+                
+                # Update previous response with game state, move and thought
+                prev_response = f"game_state: {game_state}\nmove: {move}\nthought: {thought}"
+
+            print("[debug] previous response:")
+            print(prev_response)
             elapsed_time = time.time() - start_time
-            time.sleep(1)
-            print("[debug] previous message:")
-            print("\n".join(prev_responses))
-
-            if key == "r":
-                evidence = True
-            if key == "b":
-                evidence = False
-
+            time.sleep(3)
+            print(f"[INFO] Move executed in {elapsed_time:.2f} seconds.")
     except KeyboardInterrupt:
-        print("\n[INFO] Stopping game automation. Goodbye!")
-
+        print("\nStopped by user.")
 
 if __name__ == "__main__":
     main()
