+++ tools/serving/api_providers.py
@@ -304,91 +304,6 @@ def deepseek_text_reasoning_completion(system_prompt, model_name, prompt):
     return content
     
 
-def anthropic_completion(system_prompt, model_name, base64_image, prompt, thinking=False):
-    print(f"anthropic vision-text activated... thinking: {thinking}")
-    client = anthropic.Anthropic(api_key=os.getenv("ANTHROPIC_API_KEY"))
-    messages = [
-                {
-                    "role": "user",
-                    "content": [
-                        {
-                            "type": "image",
-                            "source": {
-                                "type": "base64",
-                                "media_type": "image/png",
-                                "data": base64_image,
-                            },
-                        },
-                        {
-                            "type": "text",
-                            "text": prompt
-                        },
-                    ],
-                }
-            ]
-    if thinking:
-        with client.messages.stream(
-                max_tokens=20000,
-                thinking={
-                    "type": "enabled",
-                    "budget_tokens": 16000
-                },
-                messages=messages,
-                temperature=1,
-                system=system_prompt,
-                model=model_name, # claude-3-5-sonnet-20241022 # claude-3-7-sonnet-20250219
-            ) as stream:
-                partial_chunks = []
-                for chunk in stream.text_stream:
-                    partial_chunks.append(chunk)
-    else:
-         
-        with client.messages.stream(
-                max_tokens=1024,
-                messages=messages,
-                temperature=0,
-                system=system_prompt,
-                model=model_name, # claude-3-5-sonnet-20241022 # claude-3-7-sonnet-20250219
-            ) as stream:
-                partial_chunks = []
-                for chunk in stream.text_stream:
-                    partial_chunks.append(chunk)
-        
-    generated_code_str = "".join(partial_chunks)
-    
-    print(f"model name: {model_name}")
-    messages = [
-        {
-            "role": "user",
-            "content": [
-                {
-                    "type": "text",
-                    "text": prompt
-                },
-            ],
-        }
-    ]
-
-    # Determine correct token parameter
-    token_param = "max_completion_tokens" if "o3-mini" in model_name else "max_tokens"
-    
-    # Prepare request parameters dynamically
-    request_params = {
-        "model": model_name,
-        "messages": messages,
-        token_param: 100000,
-        "reasoning_effort": "medium"
-    }
-    
-    # Only add 'temperature' if the model supports it
-    if "o3-mini" not in model_name:  # Assuming o3-mini doesn't support 'temperature'
-        request_params["temperature"] = temperature
-
-    response = client.chat.completions.create(**request_params)
-
-    generated_str = response.choices[0].message.content
-     
-    return generated_str
 
 def openai_multiimage_completion(system_prompt, model_name, prompt, list_content, list_image_base64):
     client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
